Add `handleCommentSubmit` function to submit data
... and update the comment list *iff* successful.

Note that we're passing a callback `onCommentSubmit` as a prop into
the `CommentForm`. This allows the `CommentForm` to push the new data to
the `CommentBox`, which can then update itself with the new comment.

Beware that this doesn't actually work if you're using a static
`comments.json` file because the webserver won't accept `POST` requests.